+++ Source/DafnyCore/AST/Expressions/Expressions.cs
@@ -825,80 +825,6 @@ public abstract class Expression : INode {
   public virtual IEnumerable<INode> Children => SubExpressions;
 }
 
-/// <summary>
-/// Instances of this class are introduced during resolution to indicate that a static method or function has
-/// been invoked without specifying a receiver (that is, by just giving the name of the enclosing class).
-/// </summary>
-public class StaticReceiverExpr : LiteralExpr {
-  public readonly Type UnresolvedType;
-  private bool Implicit;
-  public Expression OriginalResolved;
-
-  public StaticReceiverExpr(IToken tok, Type t, bool isImplicit)
-    : base(tok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(t != null);
-    UnresolvedType = t;
-    Implicit = isImplicit;
-    OriginalResolved = null;
-  }
-
-  /// <summary>
-  /// Constructs a resolved LiteralExpr representing the fictitious static-receiver literal whose type is
-  /// "cl" parameterized by the type arguments of "cl" itself.
-  /// </summary>
-  public StaticReceiverExpr(IToken tok, TopLevelDeclWithMembers cl, bool isImplicit)
-    : base(tok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(cl != null);
-    var typeArgs = cl.TypeArgs.ConvertAll(tp => (Type)new UserDefinedType(tp));
-    Type = new UserDefinedType(tok, cl is ClassDecl klass && klass.IsDefaultClass ? cl.Name : cl.Name + "?", cl, typeArgs);
-    UnresolvedType = Type;
-    Implicit = isImplicit;
-  }
-
-  /// <summary>
-  /// Constructs a resolved LiteralExpr representing the fictitious literal whose type is
-  /// "cl" parameterized according to the type arguments to "t".  It is assumed that "t" denotes
-  /// a class or trait that (possibly reflexively or transitively) extends "cl".
-  /// Examples:
-  /// * If "t" denotes "C(G)" and "cl" denotes "C", then the type of the StaticReceiverExpr
-  ///   will be "C(G)".
-  /// * Suppose "C" is a class that extends a trait "T"; then, if "t" denotes "C" and "cl" denotes
-  ///   "T", then the type of the StaticReceiverExpr will be "T".
-  /// * Suppose "C(X)" is a class that extends "T(f(X))", and that "T(Y)" is
-  ///   a trait that in turn extends trait "W(g(Y))".  If "t" denotes type "C(G)" and "cl" denotes "W",
-  ///   then type of the StaticReceiverExpr will be "T(g(f(G)))".
-  /// </summary>
-  public StaticReceiverExpr(IToken tok, UserDefinedType t, TopLevelDeclWithMembers cl, bool isImplicit, Expression lhs = null)
-    : base(tok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(t.ResolvedClass != null);
-    Contract.Requires(cl != null);
-    var top = t.AsTopLevelTypeWithMembersBypassInternalSynonym;
-    if (top != cl) {
-      Contract.Assert(top != null);
-      var clArgsInTermsOfTFormals = cl.TypeArgs.ConvertAll(tp => top.ParentFormalTypeParametersToActuals[tp]);
-      var subst = TypeParameter.SubstitutionMap(top.TypeArgs, t.TypeArgs);
-      var typeArgs = clArgsInTermsOfTFormals.ConvertAll(ty => ty.Subst(subst));
-      Type = new UserDefinedType(tok, cl.Name, cl, typeArgs);
-    } else if (t.Name != cl.Name) {  // t may be using the name "C?", and we'd prefer it read "C"
-      Type = new UserDefinedType(tok, cl.Name, cl, t.TypeArgs);
-    } else {
-      Type = t;
-    }
-    UnresolvedType = Type;
-    Implicit = isImplicit;
-    OriginalResolved = lhs;
-  }
-
-  public override bool IsImplicit {
-    get { return Implicit; }
-  }
-
-  public override IEnumerable<INode> Children => base.Children.Concat(Type.Nodes);
-}
-
 public class LiteralExpr : Expression {
   /// <summary>
   /// One of the following:
@@ -1292,234 +1218,6 @@ public class SeqDisplayExpr : DisplayExpression {
   }
 }
 
-public class MemberSelectExpr : Expression, IHasUsages {
-  public readonly Expression Obj;
-  public string MemberName;
-  [FilledInDuringResolution] public MemberDecl Member;    // will be a Field or Function
-  [FilledInDuringResolution] public Label /*?*/ AtLabel;  // non-null for a two-state selection
-  [FilledInDuringResolution] public bool InCompiledContext;
-
-  /// <summary>
-  /// TypeApplication_AtEnclosingClass is the list of type arguments used to instantiate the type that
-  /// declares Member (which is some supertype of the receiver type).
-  /// </summary>
-  [FilledInDuringResolution] public List<Type> TypeApplication_AtEnclosingClass;
-
-  /// <summary>
-  ///  TypeApplication_JustMember is the list of type arguments used to instantiate the type parameters
-  /// of Member.
-  /// </summary>
-  [FilledInDuringResolution] public List<Type> TypeApplication_JustMember;
-
-  /// <summary>
-  /// Returns a mapping from formal type parameters to actual type arguments. For example, given
-  ///     trait T<A> {
-  ///       function F<X>(): bv8 { ... }
-  ///     }
-  ///     class C<B, D> extends T<map<B, D>> { }
-  /// and MemberSelectExpr o.F<int> where o has type C<real, bool>, the type map returned is
-  ///     A -> map<real, bool>
-  ///     X -> int
-  /// To also include B and D in the mapping, use TypeArgumentSubstitutionsWithParents instead.
-  /// </summary>
-  public Dictionary<TypeParameter, Type> TypeArgumentSubstitutionsAtMemberDeclaration() {
-    Contract.Requires(WasResolved());
-    Contract.Ensures(Contract.Result<Dictionary<TypeParameter, Type>>() != null);
-
-    var subst = new Dictionary<TypeParameter, Type>();
-
-    // Add the mappings from the member's own type parameters
-    if (Member is ICallable icallable) {
-      Contract.Assert(TypeApplication_JustMember.Count == icallable.TypeArgs.Count);
-      for (var i = 0; i < icallable.TypeArgs.Count; i++) {
-        subst.Add(icallable.TypeArgs[i], TypeApplication_JustMember[i]);
-      }
-    } else {
-      Contract.Assert(TypeApplication_JustMember.Count == 0);
-    }
-
-    // Add the mappings from the enclosing class.
-    TopLevelDecl cl = Member.EnclosingClass;
-    // Expand the type down to its non-null type, if any
-    if (cl != null) {
-      Contract.Assert(cl.TypeArgs.Count == TypeApplication_AtEnclosingClass.Count);
-      for (var i = 0; i < cl.TypeArgs.Count; i++) {
-        subst.Add(cl.TypeArgs[i], TypeApplication_AtEnclosingClass[i]);
-      }
-    }
-
-    return subst;
-  }
-
-  /// <summary>
-  /// Returns a mapping from formal type parameters to actual type arguments. For example, given
-  ///     trait T<A> {
-  ///       function F<X>(): bv8 { ... }
-  ///     }
-  ///     class C<B, D> extends T<map<B, D>> { }
-  /// and MemberSelectExpr o.F<int> where o has type C<real, bool>, the type map returned is
-  ///     A -> map<real, bool>
-  ///     B -> real
-  ///     D -> bool
-  ///     X -> int
-  /// NOTE: This method should be called only when all types have been fully and successfully
-  /// resolved. During type inference, when there may still be some unresolved proxies, use
-  /// TypeArgumentSubstitutionsAtMemberDeclaration instead.
-  /// </summary>
-  public Dictionary<TypeParameter, Type> TypeArgumentSubstitutionsWithParents() {
-    Contract.Requires(WasResolved());
-    Contract.Ensures(Contract.Result<Dictionary<TypeParameter, Type>>() != null);
-
-    return TypeArgumentSubstitutionsWithParentsAux(Obj.Type, Member, TypeApplication_JustMember);
-  }
-
-  public static Dictionary<TypeParameter, Type> TypeArgumentSubstitutionsWithParentsAux(Type receiverType, MemberDecl member, List<Type> typeApplicationMember) {
-    Contract.Requires(receiverType != null);
-    Contract.Requires(member != null);
-    Contract.Requires(typeApplicationMember != null);
-    Contract.Ensures(Contract.Result<Dictionary<TypeParameter, Type>>() != null);
-
-    var subst = new Dictionary<TypeParameter, Type>();
-
-    // Add the mappings from the member's own type parameters
-    if (member is ICallable) {
-      // Make sure to include the member's type parameters all the way up the inheritance chain
-      for (var ancestor = member; ancestor != null; ancestor = ancestor.OverriddenMember) {
-        var icallable = (ICallable)ancestor;
-        Contract.Assert(typeApplicationMember.Count == icallable.TypeArgs.Count);
-        for (var i = 0; i < icallable.TypeArgs.Count; i++) {
-          subst.Add(icallable.TypeArgs[i], typeApplicationMember[i]);
-        }
-      }
-    } else {
-      Contract.Assert(typeApplicationMember.Count == 0);
-    }
-
-    // Add the mappings from the receiver's type "cl"
-    var udt = receiverType.NormalizeExpand() as UserDefinedType;
-    if (udt != null) {
-      if (udt.ResolvedClass is InternalTypeSynonymDecl isyn) {
-        udt = isyn.RhsWithArgumentIgnoringScope(udt.TypeArgs) as UserDefinedType;
-      }
-      if (udt.ResolvedClass is NonNullTypeDecl nntd) {
-        udt = nntd.RhsWithArgumentIgnoringScope(udt.TypeArgs) as UserDefinedType;
-      }
-    }
-    var cl = udt?.ResolvedClass;
-
-    if (cl != null) {
-      Contract.Assert(cl.TypeArgs.Count == udt.TypeArgs.Count);
-      for (var i = 0; i < cl.TypeArgs.Count; i++) {
-        subst.Add(cl.TypeArgs[i], udt.TypeArgs[i]);
-      }
-
-      // Add in the mappings from parent types' formal type parameters to types
-      if (cl is TopLevelDeclWithMembers cls) {
-        foreach (var entry in cls.ParentFormalTypeParametersToActuals) {
-          var v = entry.Value.Subst(subst);
-          subst.Add(entry.Key, v);
-        }
-      }
-    }
-
-    return subst;
-  }
-
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(Obj != null);
-    Contract.Invariant(MemberName != null);
-    Contract.Invariant((Member != null) == (TypeApplication_AtEnclosingClass != null));  // TypeApplication_* are set whenever Member is set
-    Contract.Invariant((Member != null) == (TypeApplication_JustMember != null));  // TypeApplication_* are set whenever Member is set
-  }
-
-  public MemberSelectExpr(IToken tok, Expression obj, string memberName)
-    : base(tok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(obj != null);
-    Contract.Requires(memberName != null);
-    this.Obj = obj;
-    this.MemberName = memberName;
-  }
-
-  /// <summary>
-  /// Returns a resolved MemberSelectExpr for a field.
-  /// </summary>
-  public MemberSelectExpr(IToken tok, Expression obj, Field field)
-    : this(tok, obj, field.Name) {
-    Contract.Requires(tok != null);
-    Contract.Requires(obj != null);
-    Contract.Requires(field != null);
-    Contract.Requires(obj.Type != null);  // "obj" is required to be resolved
-
-    this.Member = field;  // resolve here
-
-    var receiverType = obj.Type.NormalizeExpand();
-    this.TypeApplication_AtEnclosingClass = receiverType.TypeArgs;
-    this.TypeApplication_JustMember = new List<Type>();
-    this.ResolvedOutparameterTypes = new List<Type>();
-
-    var typeMap = new Dictionary<TypeParameter, Type>();
-    if (receiverType is UserDefinedType udt) {
-      var cl = udt.ResolvedClass as TopLevelDeclWithMembers;
-      Contract.Assert(cl != null);
-      Contract.Assert(cl.TypeArgs.Count == TypeApplication_AtEnclosingClass.Count);
-      for (var i = 0; i < cl.TypeArgs.Count; i++) {
-        typeMap.Add(cl.TypeArgs[i], TypeApplication_AtEnclosingClass[i]);
-      }
-      foreach (var entry in cl.ParentFormalTypeParametersToActuals) {
-        var v = entry.Value.Subst(typeMap);
-        typeMap.Add(entry.Key, v);
-      }
-    } else if (field.EnclosingClass == null) {
-      // leave typeMap as the empty substitution
-    } else {
-      Contract.Assert(field.EnclosingClass.TypeArgs.Count == TypeApplication_AtEnclosingClass.Count);
-      for (var i = 0; i < field.EnclosingClass.TypeArgs.Count; i++) {
-        typeMap.Add(field.EnclosingClass.TypeArgs[i], TypeApplication_AtEnclosingClass[i]);
-      }
-    }
-    this.Type = field.Type.Subst(typeMap);  // resolve here
-  }
-
-  public void MemberSelectCase(Action<Field> fieldK, Action<Function> functionK) {
-    MemberSelectCase<bool>(
-      f => {
-        fieldK(f);
-        return true;
-      },
-      f => {
-        functionK(f);
-        return true;
-      });
-  }
-
-  public A MemberSelectCase<A>(Func<Field, A> fieldK, Func<Function, A> functionK) {
-    var field = Member as Field;
-    var function = Member as Function;
-    if (field != null) {
-      return fieldK(field);
-    } else {
-      Contract.Assert(function != null);
-      return functionK(function);
-    }
-  }
-
-  public override IEnumerable<Expression> SubExpressions {
-    get { yield return Obj; }
-  }
-
-  public override IEnumerable<Type> ComponentTypes => Util.Concat(TypeApplication_AtEnclosingClass, TypeApplication_JustMember);
-
-  [FilledInDuringResolution] public List<Type> ResolvedOutparameterTypes;
-
-  public IEnumerable<IDeclarationOrUsage> GetResolvedDeclarations() {
-    return new[] { Member };
-  }
-
-  public IToken NameToken => tok;
-}
-
 public class SeqSelectExpr : Expression {
   public readonly bool SelectOne;  // false means select a range
   public readonly Expression Seq;
@@ -1835,58 +1533,6 @@ public class MultiSetFormingExpr : Expression {
 
   public override IEnumerable<Expression> SubExpressions {
     get { yield return E; }
-  }
-}
-
-public class OldExpr : Expression {
-  [Peer]
-  public readonly Expression E;
-  public readonly string/*?*/ At;
-  [FilledInDuringResolution] public bool Useless = false;
-  [FilledInDuringResolution] public Label/*?*/ AtLabel;  // after that, At==null iff AtLabel==null
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(E != null);
-  }
-
-  [Captured]
-  public OldExpr(IToken tok, Expression expr, string at = null)
-    : base(tok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(expr != null);
-    cce.Owner.AssignSame(this, expr);
-    E = expr;
-    At = at;
-  }
-
-  public override IEnumerable<Expression> SubExpressions {
-    get { yield return E; }
-  }
-}
-
-public class UnchangedExpr : Expression {
-  public readonly List<FrameExpression> Frame;
-  public readonly string/*?*/ At;
-  [FilledInDuringResolution] public Label/*?*/ AtLabel;  // after that, At==null iff AtLabel==null
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(Frame != null);
-  }
-
-  public UnchangedExpr(IToken tok, List<FrameExpression> frame, string/*?*/ at)
-    : base(tok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(frame != null);
-    this.Frame = frame;
-    this.At = at;
-  }
-
-  public override IEnumerable<Expression> SubExpressions {
-    get {
-      foreach (var fe in Frame) {
-        yield return fe.E;
-      }
-    }
   }
 }
 
@@ -2424,71 +2070,6 @@ public class TernaryExpr : Expression {
   }
 }
 
-public class LetExpr : Expression, IAttributeBearingDeclaration, IBoundVarsBearingExpression {
-  public readonly List<CasePattern<BoundVar>> LHSs;
-  public readonly List<Expression> RHSs;
-  public readonly Expression Body;
-  public readonly bool Exact;  // Exact==true means a regular let expression; Exact==false means an assign-such-that expression
-  public readonly Attributes Attributes;
-  Attributes IAttributeBearingDeclaration.Attributes => Attributes;
-  [FilledInDuringResolution] public List<ComprehensionExpr.BoundedPool> Constraint_Bounds;  // null for Exact=true and for when expression is in a ghost context
-  // invariant Constraint_Bounds == null || Constraint_Bounds.Count == BoundVars.Count;
-  private Expression translationDesugaring;  // filled in during translation, lazily; to be accessed only via Translation.LetDesugaring; always null when Exact==true
-  private Translator lastTranslatorUsed; // avoid clashing desugaring between translators
-
-  public IToken BodyStartTok = Token.NoToken;
-  public IToken BodyEndTok = Token.NoToken;
-  IToken IRegion.BodyStartTok { get { return BodyStartTok; } }
-  IToken IRegion.BodyEndTok { get { return BodyEndTok; } }
-
-  public void setTranslationDesugaring(Translator trans, Expression expr) {
-    lastTranslatorUsed = trans;
-    translationDesugaring = expr;
-  }
-
-  public Expression getTranslationDesugaring(Translator trans) {
-    if (lastTranslatorUsed == trans) {
-      return translationDesugaring;
-    } else {
-      return null;
-    }
-  }
-
-  public LetExpr(IToken tok, List<CasePattern<BoundVar>> lhss, List<Expression> rhss, Expression body, bool exact, Attributes attrs = null)
-    : base(tok) {
-    LHSs = lhss;
-    RHSs = rhss;
-    Body = body;
-    Exact = exact;
-    Attributes = attrs;
-  }
-  public override IEnumerable<Expression> SubExpressions {
-    get {
-      foreach (var e in Attributes.SubExpressions(Attributes)) {
-        yield return e;
-      }
-      foreach (var rhs in RHSs) {
-        yield return rhs;
-      }
-      yield return Body;
-    }
-  }
-
-  public override IEnumerable<Type> ComponentTypes => BoundVars.Select(bv => bv.Type);
-
-  public IEnumerable<BoundVar> BoundVars {
-    get {
-      foreach (var lhs in LHSs) {
-        foreach (var bv in lhs.Vars) {
-          yield return bv;
-        }
-      }
-    }
-  }
-
-  public IEnumerable<BoundVar> AllBoundVars => BoundVars;
-}
-
 public class LetOrFailExpr : ConcreteSyntaxExpression {
   public readonly CasePattern<BoundVar>/*?*/ Lhs; // null means void-error handling: ":- E; F", non-null means "var pat :- E; F"
   public readonly Expression Rhs;
@@ -2498,458 +2079,6 @@ public class LetOrFailExpr : ConcreteSyntaxExpression {
     Lhs = lhs;
     Rhs = rhs;
     Body = body;
-  }
-}
-
-/// <summary>
-/// A ComprehensionExpr has the form:
-///   BINDER { x [: Type] [<- Domain] [Attributes] [| Range] } [:: Term(x)]
-/// Where BINDER is currently "forall", "exists", "iset"/"set", or "imap"/"map".
-///
-/// Quantifications used to only support a single range, but now each
-/// quantified variable can have a range attached.
-/// The overall Range is now filled in by the parser by extracting any implicit
-/// "x in Domain" constraints and per-variable Range constraints into a single conjunct.
-///
-/// The Term is optional if the expression only has one quantified variable,
-/// but required otherwise.
-///
-/// LambdaExpr also inherits from this base class but isn't really a comprehension,
-/// and should be considered implementation inheritance.
-/// </summary>
-public abstract class ComprehensionExpr : Expression, IAttributeBearingDeclaration, IBoundVarsBearingExpression {
-  public virtual string WhatKind => "comprehension";
-  public readonly List<BoundVar> BoundVars;
-  public readonly Expression Range;
-  private Expression term;
-  public Expression Term { get { return term; } }
-  public IEnumerable<BoundVar> AllBoundVars => BoundVars;
-
-  public IToken BodyStartTok = Token.NoToken;
-  public IToken BodyEndTok = Token.NoToken;
-  IToken IRegion.BodyStartTok { get { return BodyStartTok; } }
-  IToken IRegion.BodyEndTok { get { return BodyEndTok; } }
-
-  public void UpdateTerm(Expression newTerm) {
-    term = newTerm;
-  }
-
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(BoundVars != null);
-    Contract.Invariant(Term != null);
-  }
-
-  public Attributes Attributes;
-  Attributes IAttributeBearingDeclaration.Attributes => Attributes;
-
-  public abstract class BoundedPool {
-    [Flags]
-    public enum PoolVirtues { None = 0, Finite = 1, Enumerable = 2, IndependentOfAlloc = 4, IndependentOfAlloc_or_ExplicitAlloc = 8 }
-    public abstract PoolVirtues Virtues { get; }
-    /// <summary>
-    /// A higher preference is better.
-    /// A preference below 2 is a last-resort bounded pool. Bounds discovery will not consider
-    /// such a pool to be final until there are no other choices.
-    ///
-    /// For easy reference, here is the BoundedPool hierarchy and their preference levels:
-    ///
-    /// 0: AllocFreeBoundedPool
-    /// 0: ExplicitAllocatedBoundedPool
-    /// 0: SpecialAllocIndependenceAllocatedBoundedPool
-    /// 0: OlderBoundedPool
-    ///
-    /// 1: WiggleWaggleBound
-    ///
-    /// 2: SuperSetBoundedPool
-    /// 2: DatatypeInclusionBoundedPool
-    ///
-    /// 3: SubSetBoundedPool
-    ///
-    /// 4: IntBoundedPool with one bound
-    /// 5: IntBoundedPool with both bounds
-    /// 5: CharBoundedPool
-    ///
-    /// 8: DatatypeBoundedPool
-    ///
-    /// 10: CollectionBoundedPool
-    ///     - SetBoundedPool
-    ///     - MapBoundedPool
-    ///     - SeqBoundedPool
-    ///
-    /// 14: BoolBoundedPool
-    ///
-    /// 15: ExactBoundedPool
-    /// </summary>
-    public abstract int Preference(); // higher is better
-
-    public static BoundedPool GetBest(List<BoundedPool> bounds, PoolVirtues requiredVirtues) {
-      Contract.Requires(bounds != null);
-      bounds = CombineIntegerBounds(bounds);
-      BoundedPool best = null;
-      foreach (var bound in bounds) {
-        if ((bound.Virtues & requiredVirtues) == requiredVirtues) {
-          if (best == null || bound.Preference() > best.Preference()) {
-            best = bound;
-          }
-        }
-      }
-      return best;
-    }
-    public static List<VT> MissingBounds<VT>(List<VT> vars, List<BoundedPool> bounds, PoolVirtues requiredVirtues = PoolVirtues.None) where VT : IVariable {
-      Contract.Requires(vars != null);
-      Contract.Requires(bounds == null || vars.Count == bounds.Count);
-      Contract.Ensures(Contract.Result<List<VT>>() != null);
-      var missing = new List<VT>();
-      for (var i = 0; i < vars.Count; i++) {
-        if (bounds == null || bounds[i] == null || (bounds[i].Virtues & requiredVirtues) != requiredVirtues) {
-          missing.Add(vars[i]);
-        }
-      }
-      return missing;
-    }
-    public static List<bool> HasBounds(List<BoundedPool> bounds, PoolVirtues requiredVirtues = PoolVirtues.None) {
-      Contract.Requires(bounds != null);
-      Contract.Ensures(Contract.Result<List<bool>>() != null);
-      Contract.Ensures(Contract.Result<List<bool>>().Count == bounds.Count);
-      return bounds.ConvertAll(bound => bound != null && (bound.Virtues & requiredVirtues) == requiredVirtues);
-    }
-    static List<BoundedPool> CombineIntegerBounds(List<BoundedPool> bounds) {
-      var lowerBounds = new List<IntBoundedPool>();
-      var upperBounds = new List<IntBoundedPool>();
-      var others = new List<BoundedPool>();
-      foreach (var b in bounds) {
-        var ib = b as IntBoundedPool;
-        if (ib != null && ib.UpperBound == null) {
-          lowerBounds.Add(ib);
-        } else if (ib != null && ib.LowerBound == null) {
-          upperBounds.Add(ib);
-        } else {
-          others.Add(b);
-        }
-      }
-      // pair up the bounds
-      var n = Math.Min(lowerBounds.Count, upperBounds.Count);
-      for (var i = 0; i < n; i++) {
-        others.Add(new IntBoundedPool(lowerBounds[i].LowerBound, upperBounds[i].UpperBound));
-      }
-      for (var i = n; i < lowerBounds.Count; i++) {
-        others.Add(lowerBounds[i]);
-      }
-      for (var i = n; i < upperBounds.Count; i++) {
-        others.Add(upperBounds[i]);
-      }
-      return others;
-    }
-  }
-  public class ExactBoundedPool : BoundedPool {
-    public readonly Expression E;
-    public ExactBoundedPool(Expression e) {
-      Contract.Requires(e != null);
-      E = e;
-    }
-    public override PoolVirtues Virtues => PoolVirtues.Finite | PoolVirtues.Enumerable | PoolVirtues.IndependentOfAlloc | PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc;
-    public override int Preference() => 15;  // the best of all bounds
-  }
-  public class BoolBoundedPool : BoundedPool {
-    public override PoolVirtues Virtues => PoolVirtues.Finite | PoolVirtues.Enumerable | PoolVirtues.IndependentOfAlloc | PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc;
-    public override int Preference() => 14;
-  }
-  public class CharBoundedPool : BoundedPool {
-    public override PoolVirtues Virtues => PoolVirtues.Finite | PoolVirtues.Enumerable | PoolVirtues.IndependentOfAlloc | PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc;
-    public override int Preference() => 5;
-  }
-  public class AllocFreeBoundedPool : BoundedPool {
-    public Type Type;
-    public AllocFreeBoundedPool(Type t) {
-      Type = t;
-    }
-    public override PoolVirtues Virtues {
-      get {
-        if (Type.IsRefType) {
-          return PoolVirtues.Finite | PoolVirtues.IndependentOfAlloc | PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc;
-        } else {
-          return PoolVirtues.IndependentOfAlloc | PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc;
-        }
-      }
-    }
-    public override int Preference() => 0;
-  }
-  public class ExplicitAllocatedBoundedPool : BoundedPool {
-    public ExplicitAllocatedBoundedPool() {
-    }
-    public override PoolVirtues Virtues => PoolVirtues.Finite | PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc;
-    public override int Preference() => 0;
-  }
-  public class SpecialAllocIndependenceAllocatedBoundedPool : BoundedPool {
-    public SpecialAllocIndependenceAllocatedBoundedPool() {
-    }
-    public override PoolVirtues Virtues => PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc;
-    public override int Preference() => 0;
-  }
-  public class IntBoundedPool : BoundedPool {
-    public readonly Expression LowerBound;
-    public readonly Expression UpperBound;
-    public IntBoundedPool(Expression lowerBound, Expression upperBound) {
-      Contract.Requires(lowerBound != null || upperBound != null);
-      LowerBound = lowerBound;
-      UpperBound = upperBound;
-    }
-    public override PoolVirtues Virtues {
-      get {
-        if (LowerBound != null && UpperBound != null) {
-          return PoolVirtues.Finite | PoolVirtues.Enumerable | PoolVirtues.IndependentOfAlloc | PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc;
-        } else {
-          return PoolVirtues.Enumerable | PoolVirtues.IndependentOfAlloc | PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc;
-        }
-      }
-    }
-    public override int Preference() => LowerBound != null && UpperBound != null ? 5 : 4;
-  }
-  public abstract class CollectionBoundedPool : BoundedPool {
-    public readonly Type BoundVariableType;
-    public readonly Type CollectionElementType;
-    public readonly bool IsFiniteCollection;
-
-    public CollectionBoundedPool(Type bvType, Type collectionElementType, bool isFiniteCollection) {
-      Contract.Requires(bvType != null);
-      Contract.Requires(collectionElementType != null);
-
-      BoundVariableType = bvType;
-      CollectionElementType = collectionElementType;
-      IsFiniteCollection = isFiniteCollection;
-    }
-
-    public override PoolVirtues Virtues {
-      get {
-        var v = PoolVirtues.IndependentOfAlloc | PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc;
-        if (IsFiniteCollection) {
-          v |= PoolVirtues.Finite;
-          if (CollectionElementType.IsTestableToBe(BoundVariableType)) {
-            v |= PoolVirtues.Enumerable;
-          }
-        }
-        return v;
-      }
-    }
-    public override int Preference() => 10;
-  }
-  public class SetBoundedPool : CollectionBoundedPool {
-    public readonly Expression Set;
-
-    public SetBoundedPool(Expression set, Type bvType, Type collectionElementType, bool isFiniteCollection)
-      : base(bvType, collectionElementType, isFiniteCollection) {
-      Contract.Requires(set != null);
-      Contract.Requires(bvType != null);
-      Contract.Requires(collectionElementType != null);
-      Set = set;
-    }
-  }
-  public class SubSetBoundedPool : BoundedPool {
-    public readonly Expression UpperBound;
-    public readonly bool IsFiniteCollection;
-    public SubSetBoundedPool(Expression set, bool isFiniteCollection) {
-      UpperBound = set;
-      IsFiniteCollection = isFiniteCollection;
-    }
-    public override PoolVirtues Virtues {
-      get {
-        if (IsFiniteCollection) {
-          return PoolVirtues.Finite | PoolVirtues.Enumerable | PoolVirtues.IndependentOfAlloc | PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc;
-        } else {
-          // it's still enumerable, because at run time, all sets are finite after all
-          return PoolVirtues.Enumerable | PoolVirtues.IndependentOfAlloc | PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc;
-        }
-      }
-    }
-    public override int Preference() => 3;
-  }
-  public class SuperSetBoundedPool : BoundedPool {
-    public readonly Expression LowerBound;
-    public SuperSetBoundedPool(Expression set) { LowerBound = set; }
-    public override int Preference() => 2;
-    public override PoolVirtues Virtues {
-      get {
-        if (LowerBound.Type.MayInvolveReferences) {
-          return PoolVirtues.None;
-        } else {
-          return PoolVirtues.IndependentOfAlloc | PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc;
-        }
-      }
-    }
-  }
-  public class MultiSetBoundedPool : CollectionBoundedPool {
-    public readonly Expression MultiSet;
-
-    public MultiSetBoundedPool(Expression multiset, Type bvType, Type collectionElementType)
-      : base(bvType, collectionElementType, true) {
-      Contract.Requires(multiset != null);
-      Contract.Requires(bvType != null);
-      Contract.Requires(collectionElementType != null);
-      MultiSet = multiset;
-    }
-  }
-  public class MapBoundedPool : CollectionBoundedPool {
-    public readonly Expression Map;
-
-    public MapBoundedPool(Expression map, Type bvType, Type collectionElementType, bool isFiniteCollection)
-      : base(bvType, collectionElementType, isFiniteCollection) {
-      Contract.Requires(map != null);
-      Contract.Requires(bvType != null);
-      Contract.Requires(collectionElementType != null);
-      Map = map;
-    }
-  }
-  public class SeqBoundedPool : CollectionBoundedPool {
-    public readonly Expression Seq;
-
-    public SeqBoundedPool(Expression seq, Type bvType, Type collectionElementType)
-      : base(bvType, collectionElementType, true) {
-      Contract.Requires(seq != null);
-      Contract.Requires(bvType != null);
-      Contract.Requires(collectionElementType != null);
-      Seq = seq;
-    }
-  }
-  public class DatatypeBoundedPool : BoundedPool {
-    public readonly DatatypeDecl Decl;
-
-    public DatatypeBoundedPool(DatatypeDecl d) {
-      Contract.Requires(d != null);
-      Decl = d;
-    }
-    public override PoolVirtues Virtues => PoolVirtues.Finite | PoolVirtues.Enumerable | PoolVirtues.IndependentOfAlloc | PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc;
-    public override int Preference() => 8;
-  }
-  public class DatatypeInclusionBoundedPool : BoundedPool {
-    public readonly bool IsIndDatatype;
-    public DatatypeInclusionBoundedPool(bool isIndDatatype) : base() { IsIndDatatype = isIndDatatype; }
-    public override PoolVirtues Virtues => (IsIndDatatype ? PoolVirtues.Finite : PoolVirtues.None) | PoolVirtues.IndependentOfAlloc | PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc;
-    public override int Preference() => 2;
-  }
-
-  public class OlderBoundedPool : BoundedPool {
-    public OlderBoundedPool() {
-    }
-    public override PoolVirtues Virtues => PoolVirtues.IndependentOfAlloc | PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc;
-    public override int Preference() => 0;
-  }
-
-  [FilledInDuringResolution] public List<BoundedPool> Bounds;
-  // invariant Bounds == null || Bounds.Count == BoundVars.Count;
-
-  public List<BoundVar> UncompilableBoundVars() {
-    Contract.Ensures(Contract.Result<List<BoundVar>>() != null);
-    var v = BoundedPool.PoolVirtues.Finite | BoundedPool.PoolVirtues.Enumerable;
-    return ComprehensionExpr.BoundedPool.MissingBounds(BoundVars, Bounds, v);
-  }
-
-  public ComprehensionExpr(IToken tok, IToken endTok, List<BoundVar> bvars, Expression range, Expression term, Attributes attrs)
-    : base(tok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(cce.NonNullElements(bvars));
-    Contract.Requires(term != null);
-
-    this.BoundVars = bvars;
-    this.Range = range;
-    this.UpdateTerm(term);
-    this.Attributes = attrs;
-    this.BodyStartTok = tok;
-    this.BodyEndTok = endTok;
-  }
-
-  public override IEnumerable<Expression> SubExpressions {
-    get {
-      foreach (var e in Attributes.SubExpressions(Attributes)) {
-        yield return e;
-      }
-      if (Range != null) { yield return Range; }
-      yield return Term;
-    }
-  }
-
-  public override IEnumerable<Type> ComponentTypes => BoundVars.Select(bv => bv.Type);
-}
-
-public abstract class QuantifierExpr : ComprehensionExpr, TypeParameter.ParentType {
-  public override string WhatKind => "quantifier";
-
-  private readonly int UniqueId;
-  private static int currentQuantId = -1;
-
-  protected virtual BinaryExpr.ResolvedOpcode SplitResolvedOp { get { return BinaryExpr.ResolvedOpcode.Or; } }
-
-  private Expression SplitQuantifierToExpression() {
-    Contract.Requires(SplitQuantifier != null && SplitQuantifier.Any());
-    Expression accumulator = SplitQuantifier[0];
-    for (int tid = 1; tid < SplitQuantifier.Count; tid++) {
-      accumulator = new BinaryExpr(Term.tok, SplitResolvedOp, accumulator, SplitQuantifier[tid]);
-    }
-    return accumulator;
-  }
-
-  private List<Expression> _SplitQuantifier;
-  public List<Expression> SplitQuantifier {
-    get {
-      return _SplitQuantifier;
-    }
-    set {
-      Contract.Assert(!value.Contains(this)); // don't let it put into its own split quantifiers.
-      _SplitQuantifier = value;
-      SplitQuantifierExpression = SplitQuantifierToExpression();
-    }
-  }
-
-  internal Expression SplitQuantifierExpression { get; private set; }
-
-  static int FreshQuantId() {
-    return System.Threading.Interlocked.Increment(ref currentQuantId);
-  }
-
-  public string FullName {
-    get {
-      return "q$" + UniqueId;
-    }
-  }
-
-  public String Refresh(string prefix, FreshIdGenerator idGen) {
-    return idGen.FreshId(prefix);
-  }
-
-  public TypeParameter Refresh(TypeParameter p, FreshIdGenerator idGen) {
-    var cp = new TypeParameter(p.tok, idGen.FreshId(p.Name + "#"), p.VarianceSyntax, p.Characteristics);
-    cp.Parent = this;
-    return cp;
-  }
-  public QuantifierExpr(IToken tok, IToken endTok, List<BoundVar> bvars, Expression range, Expression term, Attributes attrs)
-    : base(tok, endTok, bvars, range, term, attrs) {
-    Contract.Requires(tok != null);
-    Contract.Requires(cce.NonNullElements(bvars));
-    Contract.Requires(term != null);
-    this.UniqueId = FreshQuantId();
-  }
-
-  public virtual Expression LogicalBody(bool bypassSplitQuantifier = false) {
-    // Don't call this on a quantifier with a Split clause: it's not a real quantifier. The only exception is the Compiler.
-    Contract.Requires(bypassSplitQuantifier || SplitQuantifier == null);
-    throw new cce.UnreachableException(); // This body is just here for the "Requires" clause
-  }
-
-  public override IEnumerable<Expression> SubExpressions {
-    get {
-      if (SplitQuantifier == null) {
-        foreach (var e in base.SubExpressions) {
-          yield return e;
-        }
-      } else {
-        foreach (var e in Attributes.SubExpressions(Attributes)) {
-          yield return e;
-        }
-        foreach (var e in SplitQuantifier) {
-          yield return e;
-        }
-      }
-    }
   }
 }
 
@@ -3630,57 +2759,6 @@ public class LitPattern : ExtendedPattern {
   public override IEnumerable<INode> Children => new[] { OrigLit };
 }
 
-public class IdPattern : ExtendedPattern, IHasUsages {
-  public bool HasParenthesis { get; }
-  public readonly String Id;
-  public readonly Type Type; // This is the syntactic type, ExtendedPatterns dissapear during resolution.
-  public List<ExtendedPattern> Arguments; // null if just an identifier; possibly empty argument list if a constructor call
-  public LiteralExpr ResolvedLit; // null if just an identifier
-  [FilledInDuringResolution]
-  public DatatypeCtor Ctor;
-
-  public bool IsWildcardPattern =>
-    Arguments == null && Id.StartsWith("_");
-
-  public void MakeAConstructor() {
-    this.Arguments = new List<ExtendedPattern>();
-  }
-
-  public IdPattern(IToken tok, String id, List<ExtendedPattern> arguments, bool isGhost = false, bool hasParenthesis = false) : base(tok, isGhost) {
-    Contract.Requires(id != null);
-    Contract.Requires(arguments != null); // Arguments can be empty, but shouldn't be null
-    HasParenthesis = hasParenthesis;
-    this.Id = id;
-    this.Type = new InferredTypeProxy();
-    this.Arguments = arguments;
-  }
-
-  public IdPattern(IToken tok, String id, Type type, List<ExtendedPattern> arguments, bool isGhost = false) : base(tok, isGhost) {
-    Contract.Requires(id != null);
-    Contract.Requires(arguments != null); // Arguments can be empty, but shouldn't be null
-    this.Id = id;
-    this.Type = type == null ? new InferredTypeProxy() : type;
-    this.Arguments = arguments;
-    this.IsGhost = isGhost;
-  }
-
-  public override string ToString() {
-    if (Arguments == null || Arguments.Count == 0) {
-      return Id;
-    } else {
-      List<string> cps = Arguments.ConvertAll<string>(x => x.ToString());
-      return string.Format("{0}({1})", Id, String.Join(", ", cps));
-    }
-  }
-
-  public override IEnumerable<INode> Children => Arguments ?? Enumerable.Empty<INode>();
-  public IEnumerable<IDeclarationOrUsage> GetResolvedDeclarations() {
-    return new IDeclarationOrUsage[] { Ctor }.Where(x => x != null);
-  }
-
-  public IToken NameToken => Tok;
-}
-
 public abstract class NestedMatchCase : INode {
   public readonly IToken Tok;
   public readonly ExtendedPattern Pat;
@@ -3693,94 +2771,6 @@ public abstract class NestedMatchCase : INode {
   }
 
   public abstract IEnumerable<INode> Children { get; }
-}
-
-public class NestedMatchCaseExpr : NestedMatchCase, IAttributeBearingDeclaration {
-  public readonly Expression Body;
-  public Attributes Attributes;
-  Attributes IAttributeBearingDeclaration.Attributes => Attributes;
-
-  public NestedMatchCaseExpr(IToken tok, ExtendedPattern pat, Expression body, Attributes attrs) : base(tok, pat) {
-    Contract.Requires(body != null);
-    this.Body = body;
-    this.Attributes = attrs;
-  }
-
-  public override IEnumerable<INode> Children => new INode[] { Body, Pat }.Concat(Attributes?.Args ?? Enumerable.Empty<INode>());
-}
-
-public class NestedMatchCaseStmt : NestedMatchCase, IAttributeBearingDeclaration {
-  public readonly List<Statement> Body;
-  public Attributes Attributes;
-  Attributes IAttributeBearingDeclaration.Attributes => Attributes;
-  public NestedMatchCaseStmt(IToken tok, ExtendedPattern pat, List<Statement> body) : base(tok, pat) {
-    Contract.Requires(body != null);
-    this.Body = body;
-    this.Attributes = null;
-  }
-  public NestedMatchCaseStmt(IToken tok, ExtendedPattern pat, List<Statement> body, Attributes attrs) : base(tok, pat) {
-    Contract.Requires(body != null);
-    this.Body = body;
-    this.Attributes = attrs;
-  }
-
-  public override IEnumerable<INode> Children => Body.Concat<INode>(Attributes?.Args ?? Enumerable.Empty<INode>());
-}
-
-public class NestedMatchStmt : ConcreteSyntaxStatement {
-  public readonly Expression Source;
-  public readonly List<NestedMatchCaseStmt> Cases;
-  public readonly bool UsesOptionalBraces;
-
-  private void InitializeAttributes() {
-    // Default case for match is false
-    bool splitMatch = Attributes.Contains(this.Attributes, "split");
-    Attributes.ContainsBool(this.Attributes, "split", ref splitMatch);
-    foreach (var c in this.Cases) {
-      if (!Attributes.Contains(c.Attributes, "split")) {
-        List<Expression> args = new List<Expression>();
-        args.Add(new LiteralExpr(c.Tok, splitMatch));
-        Attributes attrs = new Attributes("split", args, c.Attributes);
-        c.Attributes = attrs;
-      }
-    }
-  }
-
-  public override IEnumerable<Expression> NonSpecificationSubExpressions {
-    get {
-      foreach (var e in base.NonSpecificationSubExpressions) {
-        yield return e;
-      }
-      if (this.ResolvedStatement == null) {
-        yield return Source;
-      }
-    }
-  }
-  public NestedMatchStmt(IToken tok, IToken endTok, Expression source, [Captured] List<NestedMatchCaseStmt> cases, bool usesOptionalBraces, Attributes attrs = null)
-    : base(tok, endTok, attrs) {
-    Contract.Requires(source != null);
-    Contract.Requires(cce.NonNullElements(cases));
-    this.Source = source;
-    this.Cases = cases;
-    this.UsesOptionalBraces = usesOptionalBraces;
-    InitializeAttributes();
-  }
-}
-
-public class NestedMatchExpr : ConcreteSyntaxExpression {
-  public readonly Expression Source;
-  public readonly List<NestedMatchCaseExpr> Cases;
-  public readonly bool UsesOptionalBraces;
-  public Attributes Attributes;
-
-  public NestedMatchExpr(IToken tok, Expression source, [Captured] List<NestedMatchCaseExpr> cases, bool usesOptionalBraces, Attributes attrs = null) : base(tok) {
-    Contract.Requires(source != null);
-    Contract.Requires(cce.NonNullElements(cases));
-    this.Source = source;
-    this.Cases = cases;
-    this.UsesOptionalBraces = usesOptionalBraces;
-    this.Attributes = attrs;
-  }
 }
 
 public class BoxingCastExpr : Expression {  // a BoxingCastExpr is used only as a temporary placeholding during translation
