+++ Source/DafnyCore/AST/Statements/AlternativeStmt.cs
@@ -1,0 +1,46 @@
+using System.Collections.Generic;
+using System.Diagnostics.Contracts;
+
+namespace Microsoft.Dafny;
+
+public class AlternativeStmt : Statement {
+  public readonly bool UsesOptionalBraces;
+  public readonly List<GuardedAlternative> Alternatives;
+  [ContractInvariantMethod]
+  void ObjectInvariant() {
+    Contract.Invariant(Alternatives != null);
+  }
+  public AlternativeStmt(IToken tok, IToken endTok, List<GuardedAlternative> alternatives, bool usesOptionalBraces)
+    : base(tok, endTok) {
+    Contract.Requires(tok != null);
+    Contract.Requires(endTok != null);
+    Contract.Requires(alternatives != null);
+    this.Alternatives = alternatives;
+    this.UsesOptionalBraces = usesOptionalBraces;
+  }
+  public AlternativeStmt(IToken tok, IToken endTok, List<GuardedAlternative> alternatives, bool usesOptionalBraces, Attributes attrs)
+    : base(tok, endTok, attrs) {
+    Contract.Requires(tok != null);
+    Contract.Requires(endTok != null);
+    Contract.Requires(alternatives != null);
+    this.Alternatives = alternatives;
+    this.UsesOptionalBraces = usesOptionalBraces;
+  }
+  public override IEnumerable<Statement> SubStatements {
+    get {
+      foreach (var alt in Alternatives) {
+        foreach (var s in alt.Body) {
+          yield return s;
+        }
+      }
+    }
+  }
+  public override IEnumerable<Expression> NonSpecificationSubExpressions {
+    get {
+      foreach (var e in base.NonSpecificationSubExpressions) { yield return e; }
+      foreach (var alt in Alternatives) {
+        yield return alt.Guard;
+      }
+    }
+  }
+}
