+++ Source/DafnyCore/AST/Statements/ForLoopStmt.cs
@@ -1,0 +1,38 @@
+using System.Collections.Generic;
+using System.Diagnostics.Contracts;
+
+namespace Microsoft.Dafny;
+
+public class ForLoopStmt : OneBodyLoopStmt {
+  public readonly BoundVar LoopIndex;
+  public readonly Expression Start;
+  public readonly Expression/*?*/ End;
+  public readonly bool GoingUp;
+
+  public ForLoopStmt(IToken tok, IToken endTok, BoundVar loopIndexVariable, Expression start, Expression/*?*/ end, bool goingUp,
+    List<AttributedExpression> invariants, Specification<Expression> decreases, Specification<FrameExpression> mod,
+    BlockStmt /*?*/ body, Attributes attrs)
+    : base(tok, endTok, invariants, decreases, mod, body, attrs) {
+    Contract.Requires(tok != null);
+    Contract.Requires(endTok != null);
+    Contract.Requires(loopIndexVariable != null);
+    Contract.Requires(start != null);
+    Contract.Requires(invariants != null);
+    Contract.Requires(decreases != null);
+    Contract.Requires(mod != null);
+    LoopIndex = loopIndexVariable;
+    Start = start;
+    End = end;
+    GoingUp = goingUp;
+  }
+
+  public override IEnumerable<Expression> NonSpecificationSubExpressions {
+    get {
+      foreach (var e in base.NonSpecificationSubExpressions) { yield return e; }
+      yield return Start;
+      if (End != null) {
+        yield return End;
+      }
+    }
+  }
+}
