+++ Source/DafnyCore/AST/Statements/NestedMatchCaseStmt.cs
@@ -1,0 +1,23 @@
+using System.Collections.Generic;
+using System.Diagnostics.Contracts;
+using System.Linq;
+
+namespace Microsoft.Dafny;
+
+public class NestedMatchCaseStmt : NestedMatchCase, IAttributeBearingDeclaration {
+  public readonly List<Statement> Body;
+  public Attributes Attributes;
+  Attributes IAttributeBearingDeclaration.Attributes => Attributes;
+  public NestedMatchCaseStmt(IToken tok, ExtendedPattern pat, List<Statement> body) : base(tok, pat) {
+    Contract.Requires(body != null);
+    this.Body = body;
+    this.Attributes = null;
+  }
+  public NestedMatchCaseStmt(IToken tok, ExtendedPattern pat, List<Statement> body, Attributes attrs) : base(tok, pat) {
+    Contract.Requires(body != null);
+    this.Body = body;
+    this.Attributes = attrs;
+  }
+
+  public override IEnumerable<INode> Children => Body.Concat<INode>(Attributes?.Args ?? Enumerable.Empty<INode>());
+}
