+++ Source/DafnyCore/AST/Statements/PrintStmt.cs
@@ -1,0 +1,29 @@
+using System.Collections.Generic;
+using System.Diagnostics.Contracts;
+
+namespace Microsoft.Dafny;
+
+public class PrintStmt : Statement {
+  public readonly List<Expression> Args;
+  [ContractInvariantMethod]
+  void ObjectInvariant() {
+    Contract.Invariant(cce.NonNullElements(Args));
+  }
+
+  public PrintStmt(IToken tok, IToken endTok, List<Expression> args)
+    : base(tok, endTok) {
+    Contract.Requires(tok != null);
+    Contract.Requires(endTok != null);
+    Contract.Requires(cce.NonNullElements(args));
+
+    Args = args;
+  }
+  public override IEnumerable<Expression> NonSpecificationSubExpressions {
+    get {
+      foreach (var e in base.NonSpecificationSubExpressions) { yield return e; }
+      foreach (var arg in Args) {
+        yield return arg;
+      }
+    }
+  }
+}
