+++ Source/DafnyCore/AST/Statements/Statements.cs
@@ -1,5 +1,3 @@
-
-using System;
 using System.Collections.Generic;
 using System.Diagnostics.Contracts;
 using System.Drawing.Imaging;
@@ -207,124 +205,6 @@ public class AssertLabel : Label {
   }
 }
 
-public abstract class PredicateStmt : Statement {
-  public readonly Expression Expr;
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(Expr != null);
-  }
-
-  public PredicateStmt(IToken tok, IToken endTok, Expression expr, Attributes attrs)
-    : base(tok, endTok, attrs) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(expr != null);
-    this.Expr = expr;
-  }
-
-  public PredicateStmt(IToken tok, IToken endTok, Expression expr)
-    : this(tok, endTok, expr, null) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(expr != null);
-    this.Expr = expr;
-  }
-}
-
-public class AssertStmt : PredicateStmt {
-  public readonly BlockStmt Proof;
-  public readonly AssertLabel Label;
-  public AssertStmt(IToken tok, IToken endTok, Expression expr, BlockStmt/*?*/ proof, AssertLabel/*?*/ label, Attributes attrs)
-    : base(tok, endTok, expr, attrs) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(expr != null);
-    Proof = proof;
-    Label = label;
-  }
-  public override IEnumerable<Statement> SubStatements {
-    get {
-      if (Proof != null) {
-        yield return Proof;
-      }
-    }
-  }
-  public void AddCustomizedErrorMessage(IToken tok, string s) {
-    var args = new List<Expression>() { new StringLiteralExpr(tok, s, true) };
-    IToken openBrace = tok;
-    IToken closeBrace = new Token(tok.line, tok.col + 7 + s.Length + 1); // where 7 = length(":error ")
-    this.Attributes = new UserSuppliedAttributes(tok, openBrace, closeBrace, args, this.Attributes);
-  }
-
-  public override IEnumerable<Expression> SpecificationSubExpressions {
-    get {
-      foreach (var e in base.SpecificationSubExpressions) { yield return e; }
-      yield return Expr;
-    }
-  }
-}
-
-public class ExpectStmt : PredicateStmt {
-  public Expression Message;
-  public ExpectStmt(IToken tok, IToken endTok, Expression expr, Expression message, Attributes attrs)
-    : base(tok, endTok, expr, attrs) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(expr != null);
-    this.Message = message;
-  }
-
-  public override IEnumerable<Expression> NonSpecificationSubExpressions {
-    get {
-      foreach (var e in base.NonSpecificationSubExpressions) { yield return e; }
-      yield return Expr;
-      if (Message != null) {
-        yield return Message;
-      }
-    }
-  }
-}
-
-public class AssumeStmt : PredicateStmt {
-  public AssumeStmt(IToken tok, IToken endTok, Expression expr, Attributes attrs)
-    : base(tok, endTok, expr, attrs) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(expr != null);
-  }
-  public override IEnumerable<Expression> SpecificationSubExpressions {
-    get {
-      foreach (var e in base.SpecificationSubExpressions) { yield return e; }
-      yield return Expr;
-    }
-  }
-}
-
-public class PrintStmt : Statement {
-  public readonly List<Expression> Args;
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(cce.NonNullElements(Args));
-  }
-
-  public PrintStmt(IToken tok, IToken endTok, List<Expression> args)
-    : base(tok, endTok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(cce.NonNullElements(args));
-
-    Args = args;
-  }
-  public override IEnumerable<Expression> NonSpecificationSubExpressions {
-    get {
-      foreach (var e in base.NonSpecificationSubExpressions) { yield return e; }
-      foreach (var arg in Args) {
-        yield return arg;
-      }
-    }
-  }
-}
-
 public class RevealStmt : Statement {
   public readonly List<Expression> Exprs;
   [FilledInDuringResolution] public readonly List<AssertLabel> LabeledAsserts = new List<AssertLabel>();  // to indicate that "Expr" denotes a labeled assertion
@@ -356,49 +236,6 @@ public class RevealStmt : Statement {
       return null;
     }
   }
-}
-
-/// <summary>
-/// Class "BreakStmt" represents both "break" and "continue" statements.
-/// </summary>
-public class BreakStmt : Statement, IHasUsages {
-  public readonly IToken TargetLabel;
-  public readonly bool IsContinue;
-  public string Kind => IsContinue ? "continue" : "break";
-  public readonly int BreakAndContinueCount;
-  [FilledInDuringResolution] public Statement TargetStmt;
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(TargetLabel != null || 1 <= BreakAndContinueCount);
-  }
-
-  public BreakStmt(IToken tok, IToken endTok, IToken targetLabel, bool isContinue)
-    : base(tok, endTok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(targetLabel != null);
-    this.TargetLabel = targetLabel;
-    this.IsContinue = isContinue;
-  }
-
-  /// <summary>
-  /// For "isContinue == false", represents the statement "break ^breakAndContinueCount ;".
-  /// For "isContinue == true", represents the statement "break ^(breakAndContinueCount - 1) continue;".
-  /// </summary>
-  public BreakStmt(IToken tok, IToken endTok, int breakAndContinueCount, bool isContinue)
-    : base(tok, endTok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(1 <= breakAndContinueCount);
-    this.BreakAndContinueCount = breakAndContinueCount;
-    this.IsContinue = isContinue;
-  }
-
-  public IEnumerable<IDeclarationOrUsage> GetResolvedDeclarations() {
-    return new[] { TargetStmt }.OfType<IDeclarationOrUsage>();
-  }
-
-  public IToken NameToken => Tok;
 }
 
 public abstract class ProduceStmt : Statement {
@@ -433,15 +270,6 @@ public abstract class ProduceStmt : Statement {
         }
       }
     }
-  }
-}
-
-public class ReturnStmt : ProduceStmt {
-  public bool ReverifyPost;  // set during pre-resolution refinement transformation
-  public ReturnStmt(IToken tok, IToken endTok, List<AssignmentRhs> rhss)
-    : base(tok, endTok, rhss) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
   }
 }
 
@@ -759,48 +587,6 @@ public abstract class ConcreteUpdateStatement : Statement {
 /// </summary>
 public record AttributedToken(IToken Token, Attributes Attrs) { }
 
-public class AssignSuchThatStmt : ConcreteUpdateStatement {
-  public readonly Expression Expr;
-  public readonly AttributedToken AssumeToken;
-
-  [FilledInDuringResolution] public List<ComprehensionExpr.BoundedPool> Bounds;  // null for a ghost statement
-  // invariant Bounds == null || Bounds.Count == BoundVars.Count;
-  [FilledInDuringResolution] public List<IVariable> MissingBounds;  // remains "null" if bounds can be found
-  // invariant Bounds == null || MissingBounds == null;
-  public class WiggleWaggleBound : ComprehensionExpr.BoundedPool {
-    public override PoolVirtues Virtues => PoolVirtues.Enumerable | PoolVirtues.IndependentOfAlloc | PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc;
-    public override int Preference() => 1;
-  }
-
-  public override IEnumerable<INode> Children => Lhss.Concat<INode>(new[] { Expr });
-
-  /// <summary>
-  /// "assumeToken" is allowed to be "null", in which case the verifier will check that a RHS value exists.
-  /// If "assumeToken" is non-null, then it should denote the "assume" keyword used in the statement.
-  /// </summary>
-  public AssignSuchThatStmt(IToken tok, IToken endTok, List<Expression> lhss, Expression expr, AttributedToken assumeToken, Attributes attrs)
-    : base(tok, endTok, lhss, attrs) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(cce.NonNullElements(lhss));
-    Contract.Requires(lhss.Count != 0);
-    Contract.Requires(expr != null);
-    Expr = expr;
-    if (assumeToken != null) {
-      AssumeToken = assumeToken;
-    }
-  }
-  public override IEnumerable<Expression> NonSpecificationSubExpressions {
-    get {
-      foreach (var e in base.NonSpecificationSubExpressions) { yield return e; }
-      yield return Expr;
-      foreach (var lhs in Lhss) {
-        yield return lhs;
-      }
-    }
-  }
-}
-
 public class UpdateStmt : ConcreteUpdateStatement {
   public readonly List<AssignmentRhs> Rhss;
   public readonly bool CanMutateKnownState;
@@ -838,130 +624,6 @@ public class UpdateStmt : ConcreteUpdateStatement {
     Contract.Requires(lhss.Count != 0 || rhss.Count == 1);
     Rhss = rhss;
     CanMutateKnownState = mutate;
-  }
-}
-
-public class AssignOrReturnStmt : ConcreteUpdateStatement {
-  public readonly Expression Rhs; // this is the unresolved RHS, and thus can also be a method call
-  public readonly List<AssignmentRhs> Rhss;
-  public readonly AttributedToken KeywordToken;
-  [FilledInDuringResolution] public readonly List<Statement> ResolvedStatements = new List<Statement>();
-  public override IEnumerable<Statement> SubStatements {
-    get { return ResolvedStatements; }
-  }
-
-  public override IEnumerable<INode> Children => ResolvedStatements;
-
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(Lhss != null);
-    Contract.Invariant(
-      Lhss.Count == 0 ||                   // ":- MethodOrExpresion;" which returns void success or an error
-      (Lhss.Count == 1 && Lhss[0] != null)   // "y :- MethodOrExpression;"
-    );
-    Contract.Invariant(Rhs != null);
-  }
-
-  public AssignOrReturnStmt(IToken tok, IToken endTok, List<Expression> lhss, Expression rhs, AttributedToken keywordToken, List<AssignmentRhs> rhss = null)
-    : base(tok, endTok, lhss) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(lhss != null);
-    Contract.Requires(lhss.Count <= 1);
-    Contract.Requires(rhs != null);
-    Rhs = rhs;
-    Rhss = rhss;
-    KeywordToken = keywordToken;
-  }
-}
-
-public class AssignStmt : Statement {
-  public readonly Expression Lhs;
-  public readonly AssignmentRhs Rhs;
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(Lhs != null);
-    Contract.Invariant(Rhs != null);
-  }
-
-  public AssignStmt(IToken tok, IToken endTok, Expression lhs, AssignmentRhs rhs)
-    : base(tok, endTok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(lhs != null);
-    Contract.Requires(rhs != null);
-    this.Lhs = lhs;
-    this.Rhs = rhs;
-  }
-
-  public override IEnumerable<Statement> SubStatements {
-    get {
-      foreach (var s in Rhs.SubStatements) {
-        yield return s;
-      }
-    }
-  }
-
-  public override IEnumerable<Expression> NonSpecificationSubExpressions {
-    get {
-      foreach (var e in base.NonSpecificationSubExpressions) { yield return e; }
-      yield return Lhs;
-      foreach (var ee in Rhs.SubExpressions) {
-        yield return ee;
-      }
-    }
-  }
-
-  /// <summary>
-  /// This method assumes "lhs" has been successfully resolved.
-  /// </summary>
-  public static bool LhsIsToGhost(Expression lhs) {
-    Contract.Requires(lhs != null);
-    return LhsIsToGhost_Which(lhs) == NonGhostKind.IsGhost;
-  }
-  public static bool LhsIsToGhostOrAutoGhost(Expression lhs) {
-    Contract.Requires(lhs != null);
-    return LhsIsToGhost_Which(lhs) == NonGhostKind.IsGhost || lhs.Resolved is AutoGhostIdentifierExpr;
-  }
-  public enum NonGhostKind { IsGhost, Variable, Field, ArrayElement }
-  public static string NonGhostKind_To_String(NonGhostKind gk) {
-    Contract.Requires(gk != NonGhostKind.IsGhost);
-    switch (gk) {
-      case NonGhostKind.Variable: return "non-ghost variable";
-      case NonGhostKind.Field: return "non-ghost field";
-      case NonGhostKind.ArrayElement: return "array element";
-      default:
-        Contract.Assume(false);  // unexpected NonGhostKind
-        throw new cce.UnreachableException();  // please compiler
-    }
-  }
-  /// <summary>
-  /// This method assumes "lhs" has been successfully resolved.
-  /// </summary>
-  public static NonGhostKind LhsIsToGhost_Which(Expression lhs) {
-    Contract.Requires(lhs != null);
-    lhs = lhs.Resolved;
-    if (lhs is AutoGhostIdentifierExpr) {
-      // TODO: Should we return something different for this case?
-      var x = (IdentifierExpr)lhs;
-      if (!x.Var.IsGhost) {
-        return NonGhostKind.Variable;
-      }
-    } else if (lhs is IdentifierExpr) {
-      var x = (IdentifierExpr)lhs;
-      if (!x.Var.IsGhost) {
-        return NonGhostKind.Variable;
-      }
-    } else if (lhs is MemberSelectExpr) {
-      var x = (MemberSelectExpr)lhs;
-      if (!x.Member.IsGhost) {
-        return NonGhostKind.Field;
-      }
-    } else {
-      // LHS denotes an array element, which is always non-ghost
-      return NonGhostKind.ArrayElement;
-    }
-    return NonGhostKind.IsGhost;
   }
 }
 
@@ -1120,106 +782,6 @@ public class CallStmt : Statement {
   }
 }
 
-public class BlockStmt : Statement, IRegion {
-  public readonly List<Statement> Body;
-
-  IToken IRegion.BodyStartTok => Tok;
-  IToken IRegion.BodyEndTok => EndTok;
-
-  public BlockStmt(IToken tok, IToken endTok, [Captured] List<Statement> body)
-    : base(tok, endTok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(cce.NonNullElements(body));
-    this.Body = body;
-  }
-
-  public override IEnumerable<Statement> SubStatements {
-    get { return Body; }
-  }
-
-  public virtual void AppendStmt(Statement s) {
-    Contract.Requires(s != null);
-    Body.Add(s);
-  }
-}
-
-/**
-   * Used by two phase constructors: https://dafny-lang.github.io/dafny/DafnyRef/DafnyRef#13323-two-phase-constructors
-   */
-public class DividedBlockStmt : BlockStmt {
-  public readonly List<Statement> BodyInit;  // first part of Body's statements
-  public readonly IToken SeparatorTok;  // token that separates the two parts, if any
-  public readonly List<Statement> BodyProper;  // second part of Body's statements
-  public DividedBlockStmt(IToken tok, IToken endTok, List<Statement> bodyInit, IToken/*?*/ separatorTok, List<Statement> bodyProper)
-    : base(tok, endTok, Util.Concat(bodyInit, bodyProper)) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(cce.NonNullElements(bodyInit));
-    Contract.Requires(cce.NonNullElements(bodyProper));
-    this.BodyInit = bodyInit;
-    this.SeparatorTok = separatorTok;
-    this.BodyProper = bodyProper;
-  }
-  public override void AppendStmt(Statement s) {
-    BodyProper.Add(s);
-    base.AppendStmt(s);
-  }
-}
-
-public class IfStmt : Statement {
-  public readonly bool IsBindingGuard;
-  public readonly Expression Guard;
-  public readonly BlockStmt Thn;
-  public readonly Statement Els;
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(!IsBindingGuard || (Guard is ExistsExpr && ((ExistsExpr)Guard).Range == null));
-    Contract.Invariant(Thn != null);
-    Contract.Invariant(Els == null || Els is BlockStmt || Els is IfStmt || Els is SkeletonStatement);
-  }
-  public IfStmt(IToken tok, IToken endTok, bool isBindingGuard, Expression guard, BlockStmt thn, Statement els)
-    : base(tok, endTok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(!isBindingGuard || (guard is ExistsExpr && ((ExistsExpr)guard).Range == null));
-    Contract.Requires(thn != null);
-    Contract.Requires(els == null || els is BlockStmt || els is IfStmt || els is SkeletonStatement);
-    this.IsBindingGuard = isBindingGuard;
-    this.Guard = guard;
-    this.Thn = thn;
-    this.Els = els;
-  }
-  public IfStmt(IToken tok, IToken endTok, bool isBindingGuard, Expression guard, BlockStmt thn, Statement els, Attributes attrs)
-    : base(tok, endTok, attrs) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(!isBindingGuard || (guard is ExistsExpr && ((ExistsExpr)guard).Range == null));
-    Contract.Requires(thn != null);
-    Contract.Requires(els == null || els is BlockStmt || els is IfStmt || els is SkeletonStatement);
-    this.IsBindingGuard = isBindingGuard;
-    this.Guard = guard;
-    this.Thn = thn;
-    this.Els = els;
-  }
-  public override IEnumerable<Statement> SubStatements {
-    get {
-      yield return Thn;
-      if (Els != null) {
-        yield return Els;
-      }
-    }
-  }
-  public override IEnumerable<Expression> NonSpecificationSubExpressions {
-    get {
-      foreach (var e in base.NonSpecificationSubExpressions) { yield return e; }
-      if (Guard != null) {
-        yield return Guard;
-      }
-    }
-  }
-}
-
 public class GuardedAlternative : IAttributeBearingDeclaration {
   public readonly IToken Tok;
   public readonly bool IsBindingGuard;
@@ -1256,143 +818,6 @@ public class GuardedAlternative : IAttributeBearingDeclaration {
     this.Guard = guard;
     this.Body = body;
     this.Attributes = attrs;
-  }
-}
-
-public class AlternativeStmt : Statement {
-  public readonly bool UsesOptionalBraces;
-  public readonly List<GuardedAlternative> Alternatives;
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(Alternatives != null);
-  }
-  public AlternativeStmt(IToken tok, IToken endTok, List<GuardedAlternative> alternatives, bool usesOptionalBraces)
-    : base(tok, endTok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(alternatives != null);
-    this.Alternatives = alternatives;
-    this.UsesOptionalBraces = usesOptionalBraces;
-  }
-  public AlternativeStmt(IToken tok, IToken endTok, List<GuardedAlternative> alternatives, bool usesOptionalBraces, Attributes attrs)
-    : base(tok, endTok, attrs) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(alternatives != null);
-    this.Alternatives = alternatives;
-    this.UsesOptionalBraces = usesOptionalBraces;
-  }
-  public override IEnumerable<Statement> SubStatements {
-    get {
-      foreach (var alt in Alternatives) {
-        foreach (var s in alt.Body) {
-          yield return s;
-        }
-      }
-    }
-  }
-  public override IEnumerable<Expression> NonSpecificationSubExpressions {
-    get {
-      foreach (var e in base.NonSpecificationSubExpressions) { yield return e; }
-      foreach (var alt in Alternatives) {
-        yield return alt.Guard;
-      }
-    }
-  }
-}
-
-public abstract class LoopStmt : Statement, IDeclarationOrUsage {
-  public readonly List<AttributedExpression> Invariants;
-  public readonly Specification<Expression> Decreases;
-  [FilledInDuringResolution] public bool InferredDecreases;  // says that no explicit "decreases" clause was given and an attempt was made to find one automatically (which may or may not have produced anything)
-  public readonly Specification<FrameExpression> Mod;
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(cce.NonNullElements(Invariants));
-    Contract.Invariant(Decreases != null);
-    Contract.Invariant(Mod != null);
-  }
-  public LoopStmt(IToken tok, IToken endTok, List<AttributedExpression> invariants, Specification<Expression> decreases, Specification<FrameExpression> mod)
-    : base(tok, endTok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(cce.NonNullElements(invariants));
-    Contract.Requires(decreases != null);
-    Contract.Requires(mod != null);
-
-    this.Invariants = invariants;
-    this.Decreases = decreases;
-    this.Mod = mod;
-  }
-  public LoopStmt(IToken tok, IToken endTok, List<AttributedExpression> invariants, Specification<Expression> decreases, Specification<FrameExpression> mod, Attributes attrs)
-    : base(tok, endTok, attrs) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(cce.NonNullElements(invariants));
-    Contract.Requires(decreases != null);
-    Contract.Requires(mod != null);
-
-    this.Invariants = invariants;
-    this.Decreases = decreases;
-    this.Mod = mod;
-  }
-  public IEnumerable<Expression> LoopSpecificationExpressions {
-    get {
-      foreach (var mfe in Invariants) {
-        foreach (var e in Attributes.SubExpressions(mfe.Attributes)) { yield return e; }
-        yield return mfe.E;
-      }
-      foreach (var e in Attributes.SubExpressions(Decreases.Attributes)) { yield return e; }
-      if (Decreases.Expressions != null) {
-        foreach (var e in Decreases.Expressions) {
-          yield return e;
-        }
-      }
-      foreach (var e in Attributes.SubExpressions(Mod.Attributes)) { yield return e; }
-      if (Mod.Expressions != null) {
-        foreach (var fe in Mod.Expressions) {
-          yield return fe.E;
-        }
-      }
-    }
-  }
-
-  public override IEnumerable<Expression> NonSpecificationSubExpressions {
-    get {
-      foreach (var e in base.NonSpecificationSubExpressions) {
-        yield return e;
-      }
-    }
-  }
-
-  public override IEnumerable<Expression> SpecificationSubExpressions {
-    get {
-      foreach (var e in LoopSpecificationExpressions) {
-        yield return e;
-      }
-    }
-  }
-
-  public IToken NameToken => Tok;
-}
-
-public abstract class OneBodyLoopStmt : LoopStmt {
-  public readonly BlockStmt/*?*/ Body;
-  public WhileStmt.LoopBodySurrogate/*?*/ BodySurrogate;  // set by Resolver; remains null unless Body==null
-
-  public OneBodyLoopStmt(IToken tok, IToken endTok,
-    List<AttributedExpression> invariants, Specification<Expression> decreases, Specification<FrameExpression> mod,
-    BlockStmt /*?*/ body, Attributes/*?*/ attrs)
-    : base(tok, endTok, invariants, decreases, mod, attrs) {
-    Body = body;
-  }
-
-  public override IEnumerable<Statement> SubStatements {
-    get {
-      if (Body != null) {
-        yield return Body;
-      }
-    }
   }
 }
 
@@ -1456,485 +881,6 @@ public class RefinedWhileStmt : WhileStmt {
     Contract.Requires(tok != null);
     Contract.Requires(endTok != null);
     Contract.Requires(body != null);
-  }
-}
-
-public class ForLoopStmt : OneBodyLoopStmt {
-  public readonly BoundVar LoopIndex;
-  public readonly Expression Start;
-  public readonly Expression/*?*/ End;
-  public readonly bool GoingUp;
-
-  public ForLoopStmt(IToken tok, IToken endTok, BoundVar loopIndexVariable, Expression start, Expression/*?*/ end, bool goingUp,
-    List<AttributedExpression> invariants, Specification<Expression> decreases, Specification<FrameExpression> mod,
-    BlockStmt /*?*/ body, Attributes attrs)
-    : base(tok, endTok, invariants, decreases, mod, body, attrs) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(loopIndexVariable != null);
-    Contract.Requires(start != null);
-    Contract.Requires(invariants != null);
-    Contract.Requires(decreases != null);
-    Contract.Requires(mod != null);
-    LoopIndex = loopIndexVariable;
-    Start = start;
-    End = end;
-    GoingUp = goingUp;
-  }
-
-  public override IEnumerable<Expression> NonSpecificationSubExpressions {
-    get {
-      foreach (var e in base.NonSpecificationSubExpressions) { yield return e; }
-      yield return Start;
-      if (End != null) {
-        yield return End;
-      }
-    }
-  }
-}
-
-public class AlternativeLoopStmt : LoopStmt {
-  public readonly bool UsesOptionalBraces;
-  public readonly List<GuardedAlternative> Alternatives;
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(Alternatives != null);
-  }
-  public AlternativeLoopStmt(IToken tok, IToken endTok,
-    List<AttributedExpression> invariants, Specification<Expression> decreases, Specification<FrameExpression> mod,
-    List<GuardedAlternative> alternatives, bool usesOptionalBraces)
-    : base(tok, endTok, invariants, decreases, mod) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(alternatives != null);
-    this.Alternatives = alternatives;
-    this.UsesOptionalBraces = usesOptionalBraces;
-  }
-  public AlternativeLoopStmt(IToken tok, IToken endTok,
-    List<AttributedExpression> invariants, Specification<Expression> decreases, Specification<FrameExpression> mod,
-    List<GuardedAlternative> alternatives, bool usesOptionalBraces, Attributes attrs)
-    : base(tok, endTok, invariants, decreases, mod, attrs) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(alternatives != null);
-    this.Alternatives = alternatives;
-    this.UsesOptionalBraces = usesOptionalBraces;
-  }
-  public override IEnumerable<Statement> SubStatements {
-    get {
-      foreach (var alt in Alternatives) {
-        foreach (var s in alt.Body) {
-          yield return s;
-        }
-      }
-    }
-  }
-  public override IEnumerable<Expression> NonSpecificationSubExpressions {
-    get {
-      foreach (var e in base.NonSpecificationSubExpressions) { yield return e; }
-      foreach (var alt in Alternatives) {
-        yield return alt.Guard;
-      }
-    }
-  }
-}
-
-public class ForallStmt : Statement {
-  public readonly List<BoundVar> BoundVars;  // note, can be the empty list, in which case Range denotes "true"
-  public Expression Range;  // mostly readonly, except that it may in some cases be updated during resolution to conjoin the precondition of the call in the body
-  public readonly List<AttributedExpression> Ens;
-  public readonly Statement Body;
-  public List<Expression> ForallExpressions;   // fill in by rewriter.
-  public bool CanConvert = true; //  can convert to ForallExpressions
-
-  [FilledInDuringResolution] public List<ComprehensionExpr.BoundedPool> Bounds;
-  // invariant: if successfully resolved, Bounds.Count == BoundVars.Count;
-
-  /// <summary>
-  /// Assign means there are no ensures clauses and the body consists of one update statement,
-  ///   either to an object field or to an array.
-  /// Call means there are no ensures clauses and the body consists of a single call to a (presumably
-  ///   ghost, but non-ghost is also allowed) method with no out-parameters and an empty modifies
-  ///   clause.
-  /// Proof means there is at least one ensures clause, and the body consists of any (presumably ghost,
-  ///   but non-ghost is also allowed) code without side effects on variables (including fields and array
-  ///   elements) declared outside the body itself.
-  /// Notes:
-  /// * More kinds may be allowed in the future.
-  /// * One could also allow Call to call non-ghost methods without side effects.  However, that
-  ///   would seem pointless in the program, so they are disallowed (to avoid any confusion that
-  ///   such use of the forall statement might actually have a point).
-  /// * One could allow Proof even without ensures clauses that "export" what was learned.
-  ///   However, that might give the false impression that the body is nevertheless exported.
-  /// </summary>
-  public enum BodyKind { Assign, Call, Proof }
-  [FilledInDuringResolution] public BodyKind Kind;
-
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(BoundVars != null);
-    Contract.Invariant(Range != null);
-    Contract.Invariant(BoundVars.Count != 0 || LiteralExpr.IsTrue(Range));
-    Contract.Invariant(Ens != null);
-  }
-
-  public ForallStmt(IToken tok, IToken endTok, List<BoundVar> boundVars, Attributes attrs, Expression range, List<AttributedExpression> ens, Statement body)
-    : base(tok, endTok, attrs) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(cce.NonNullElements(boundVars));
-    Contract.Requires(range != null);
-    Contract.Requires(boundVars.Count != 0 || LiteralExpr.IsTrue(range));
-    Contract.Requires(cce.NonNullElements(ens));
-    this.BoundVars = boundVars;
-    this.Range = range;
-    this.Ens = ens;
-    this.Body = body;
-  }
-
-  public Statement S0 {
-    get {
-      // dig into Body to find a single statement
-      Statement s = this.Body;
-      while (true) {
-        var block = s as BlockStmt;
-        if (block != null && block.Body.Count == 1) {
-          s = block.Body[0];
-          // dig further into s
-        } else if (s is UpdateStmt) {
-          var update = (UpdateStmt)s;
-          if (update.ResolvedStatements.Count == 1) {
-            s = update.ResolvedStatements[0];
-            // dig further into s
-          } else {
-            return s;
-          }
-        } else {
-          return s;
-        }
-      }
-    }
-  }
-
-  public override IEnumerable<Statement> SubStatements {
-    get {
-      if (Body != null) {
-        yield return Body;
-      }
-    }
-  }
-
-  public override IEnumerable<Expression> NonSpecificationSubExpressions {
-    get {
-      foreach (var e in base.NonSpecificationSubExpressions) {
-        yield return e;
-      }
-
-      yield return Range;
-    }
-  }
-  public override IEnumerable<Expression> SpecificationSubExpressions {
-    get {
-      foreach (var e in base.SpecificationSubExpressions) {
-        yield return e;
-      }
-      foreach (var ee in Ens) {
-        foreach (var e in Attributes.SubExpressions(ee.Attributes)) { yield return e; }
-        yield return ee.E;
-      }
-    }
-  }
-
-  public List<BoundVar> UncompilableBoundVars() {
-    Contract.Ensures(Contract.Result<List<BoundVar>>() != null);
-    var v = ComprehensionExpr.BoundedPool.PoolVirtues.Finite | ComprehensionExpr.BoundedPool.PoolVirtues.Enumerable;
-    return ComprehensionExpr.BoundedPool.MissingBounds(BoundVars, Bounds, v);
-  }
-}
-
-public class ModifyStmt : Statement {
-  public readonly Specification<FrameExpression> Mod;
-  public readonly BlockStmt Body;
-
-  public ModifyStmt(IToken tok, IToken endTok, List<FrameExpression> mod, Attributes attrs, BlockStmt body)
-    : base(tok, endTok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(mod != null);
-    Mod = new Specification<FrameExpression>(mod, attrs);
-    Body = body;
-  }
-
-  public override IEnumerable<Statement> SubStatements {
-    get {
-      if (Body != null) {
-        yield return Body;
-      }
-    }
-  }
-  public override IEnumerable<Expression> SpecificationSubExpressions {
-    get {
-      foreach (var e in base.SpecificationSubExpressions) { yield return e; }
-      foreach (var e in Attributes.SubExpressions(Mod.Attributes)) { yield return e; }
-      foreach (var fe in Mod.Expressions) {
-        yield return fe.E;
-      }
-    }
-  }
-}
-
-public class CalcStmt : Statement {
-  public abstract class CalcOp {
-    /// <summary>
-    /// Resulting operator "x op z" if "x this y" and "y other z".
-    /// Returns null if this and other are incompatible.
-    /// </summary>
-    [Pure]
-    public abstract CalcOp ResultOp(CalcOp other);
-
-    /// <summary>
-    /// Returns an expression "line0 this line1".
-    /// </summary>
-    [Pure]
-    public abstract Expression StepExpr(Expression line0, Expression line1);
-  }
-
-  public class BinaryCalcOp : CalcOp {
-    public readonly BinaryExpr.Opcode Op;
-
-    [ContractInvariantMethod]
-    void ObjectInvariant() {
-      Contract.Invariant(ValidOp(Op));
-    }
-
-    /// <summary>
-    /// Is op a valid calculation operator?
-    /// </summary>
-    [Pure]
-    public static bool ValidOp(BinaryExpr.Opcode op) {
-      return
-        op == BinaryExpr.Opcode.Eq || op == BinaryExpr.Opcode.Neq
-                                   || op == BinaryExpr.Opcode.Lt || op == BinaryExpr.Opcode.Le
-                                   || op == BinaryExpr.Opcode.Gt || op == BinaryExpr.Opcode.Ge
-                                   || LogicOp(op);
-    }
-
-    /// <summary>
-    /// Is op a valid operator only for Boolean lines?
-    /// </summary>
-    [Pure]
-    public static bool LogicOp(BinaryExpr.Opcode op) {
-      return op == BinaryExpr.Opcode.Iff || op == BinaryExpr.Opcode.Imp || op == BinaryExpr.Opcode.Exp;
-    }
-
-    public BinaryCalcOp(BinaryExpr.Opcode op) {
-      Contract.Requires(ValidOp(op));
-      Op = op;
-    }
-
-    /// <summary>
-    /// Does this subsume other (this . other == other . this == this)?
-    /// </summary>
-    private bool Subsumes(BinaryCalcOp other) {
-      Contract.Requires(other != null);
-      var op1 = Op;
-      var op2 = other.Op;
-      if (op1 == BinaryExpr.Opcode.Neq || op2 == BinaryExpr.Opcode.Neq) {
-        return op2 == BinaryExpr.Opcode.Eq;
-      }
-
-      if (op1 == op2) {
-        return true;
-      }
-
-      if (LogicOp(op1) || LogicOp(op2)) {
-        return op2 == BinaryExpr.Opcode.Eq ||
-               (op1 == BinaryExpr.Opcode.Imp && op2 == BinaryExpr.Opcode.Iff) ||
-               (op1 == BinaryExpr.Opcode.Exp && op2 == BinaryExpr.Opcode.Iff) ||
-               (op1 == BinaryExpr.Opcode.Eq && op2 == BinaryExpr.Opcode.Iff);
-      }
-
-      return op2 == BinaryExpr.Opcode.Eq ||
-             (op1 == BinaryExpr.Opcode.Lt && op2 == BinaryExpr.Opcode.Le) ||
-             (op1 == BinaryExpr.Opcode.Gt && op2 == BinaryExpr.Opcode.Ge);
-    }
-
-    public override CalcOp ResultOp(CalcOp other) {
-      if (other is BinaryCalcOp) {
-        var o = (BinaryCalcOp)other;
-        if (this.Subsumes(o)) {
-          return this;
-        } else if (o.Subsumes(this)) {
-          return other;
-        }
-        return null;
-      } else if (other is TernaryCalcOp) {
-        return other.ResultOp(this);
-      } else {
-        Contract.Assert(false);
-        throw new cce.UnreachableException();
-      }
-    }
-
-    public override Expression StepExpr(Expression line0, Expression line1) {
-      if (Op == BinaryExpr.Opcode.Exp) {
-        // The order of operands is reversed so that it can be turned into implication during resolution
-        return new BinaryExpr(line0.tok, Op, line1, line0);
-      } else {
-        return new BinaryExpr(line0.tok, Op, line0, line1);
-      }
-    }
-
-    public override string ToString() {
-      return BinaryExpr.OpcodeString(Op);
-    }
-
-  }
-
-  public class TernaryCalcOp : CalcOp {
-    public readonly Expression Index; // the only allowed ternary operator is ==#, so we only store the index
-
-    [ContractInvariantMethod]
-    void ObjectInvariant() {
-      Contract.Invariant(Index != null);
-    }
-
-    public TernaryCalcOp(Expression idx) {
-      Contract.Requires(idx != null);
-      Index = idx;
-    }
-
-    public override CalcOp ResultOp(CalcOp other) {
-      if (other is BinaryCalcOp) {
-        if (((BinaryCalcOp)other).Op == BinaryExpr.Opcode.Eq) {
-          return this;
-        }
-        return null;
-      } else if (other is TernaryCalcOp) {
-        var a = Index;
-        var b = ((TernaryCalcOp)other).Index;
-        var minIndex = new ITEExpr(a.tok, false, new BinaryExpr(a.tok, BinaryExpr.Opcode.Le, a, b), a, b);
-        return new TernaryCalcOp(minIndex); // ToDo: if we could compare expressions for syntactic equalty, we could use this here to optimize
-      } else {
-        Contract.Assert(false);
-        throw new cce.UnreachableException();
-      }
-    }
-
-    public override Expression StepExpr(Expression line0, Expression line1) {
-      return new TernaryExpr(line0.tok, TernaryExpr.Opcode.PrefixEqOp, Index, line0, line1);
-    }
-
-    public override string ToString() {
-      return "==#";
-    }
-
-  }
-
-  public readonly List<Expression> Lines;    // Last line is dummy, in order to form a proper step with the dangling hint
-  public readonly List<BlockStmt> Hints;     // Hints[i] comes after line i; block statement is used as a container for multiple sub-hints
-  public readonly CalcOp UserSuppliedOp;     // may be null, if omitted by the user
-  public CalcOp Op;                          // main operator of the calculation (either UserSuppliedOp or (after resolution) an inferred CalcOp)
-  public readonly List<CalcOp/*?*/> StepOps; // StepOps[i] comes after line i
-  [FilledInDuringResolution] public readonly List<Expression> Steps;    // expressions li op l<i + 1> (last step is dummy)
-  [FilledInDuringResolution] public Expression Result;                  // expression l0 ResultOp ln
-
-  public static readonly CalcOp DefaultOp = new BinaryCalcOp(BinaryExpr.Opcode.Eq);
-
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(Lines != null);
-    Contract.Invariant(cce.NonNullElements(Lines));
-    Contract.Invariant(Hints != null);
-    Contract.Invariant(cce.NonNullElements(Hints));
-    Contract.Invariant(StepOps != null);
-    Contract.Invariant(Steps != null);
-    Contract.Invariant(cce.NonNullElements(Steps));
-    Contract.Invariant(Hints.Count == Math.Max(Lines.Count - 1, 0));
-    Contract.Invariant(StepOps.Count == Hints.Count);
-  }
-
-  public CalcStmt(IToken tok, IToken endTok, CalcOp userSuppliedOp, List<Expression> lines, List<BlockStmt> hints, List<CalcOp/*?*/> stepOps, Attributes attrs)
-    : base(tok, endTok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(lines != null);
-    Contract.Requires(hints != null);
-    Contract.Requires(stepOps != null);
-    Contract.Requires(cce.NonNullElements(lines));
-    Contract.Requires(cce.NonNullElements(hints));
-    Contract.Requires(hints.Count == Math.Max(lines.Count - 1, 0));
-    Contract.Requires(stepOps.Count == hints.Count);
-    this.UserSuppliedOp = userSuppliedOp;
-    this.Lines = lines;
-    this.Hints = hints;
-    this.StepOps = stepOps;
-    this.Steps = new List<Expression>();
-    this.Result = null;
-    this.Attributes = attrs;
-  }
-
-  public override IEnumerable<Statement> SubStatements {
-    get {
-      foreach (var h in Hints) {
-        yield return h;
-      }
-    }
-  }
-  public override IEnumerable<Expression> SpecificationSubExpressions {
-    get {
-      foreach (var e in base.SpecificationSubExpressions) { yield return e; }
-      foreach (var e in Attributes.SubExpressions(Attributes)) { yield return e; }
-
-      for (int i = 0; i < Lines.Count - 1; i++) {  // note, we skip the duplicated line at the end
-        yield return Lines[i];
-      }
-      foreach (var calcop in AllCalcOps) {
-        var o3 = calcop as TernaryCalcOp;
-        if (o3 != null) {
-          yield return o3.Index;
-        }
-      }
-    }
-  }
-
-  IEnumerable<CalcOp> AllCalcOps {
-    get {
-      if (UserSuppliedOp != null) {
-        yield return UserSuppliedOp;
-      }
-      foreach (var stepop in StepOps) {
-        if (stepop != null) {
-          yield return stepop;
-        }
-      }
-    }
-  }
-
-  /// <summary>
-  /// Left-hand side of a step expression.
-  /// Note that Lhs(op.StepExpr(line0, line1)) != line0 when op is <==.
-  /// </summary>
-  public static Expression Lhs(Expression step) {
-    Contract.Requires(step is BinaryExpr || step is TernaryExpr);
-    if (step is BinaryExpr) {
-      return ((BinaryExpr)step).E0;
-    } else {
-      return ((TernaryExpr)step).E1;
-    }
-  }
-
-  /// <summary>
-  /// Right-hand side of a step expression.
-  /// Note that Rhs(op.StepExpr(line0, line1)) != line1 when op is REVERSE-IMPLICATION.
-  /// </summary>
-  public static Expression Rhs(Expression step) {
-    Contract.Requires(step is BinaryExpr || step is TernaryExpr);
-    if (step is BinaryExpr) {
-      return ((BinaryExpr)step).E1;
-    } else {
-      return ((TernaryExpr)step).E2;
-    }
   }
 }
